==> 01_Utilities.ipynb <==
# default_exp utilities

# export
import numpy as np
import ensembl_rest

# export
def getSequence(gene):
    "Get the canonical protein sequence coded by the given gene symbol"
    q = ensembl_rest.symbol_lookup(
        species='human',
        symbol=gene,params={'expand':True})
    try:
        is_canonical = [t["is_canonical"] for t in q["Transcript"]]
        t = q["Transcript"][np.argmax(is_canonical)]
        seq = ensembl_rest.sequence_id(t["Translation"]["id"])["seq"]
        return seq
    except:
        return ""

getSequence("BSCL2")
