==> 02_ProcessClinvar.ipynb <==
import pandas as pd

from ensembl_rest import sequence_id, lookup
from Bio import Entrez

Entrez.email = "zeiberg.d@northeastern.edu"

summary = pd.read_csv("/data/projects/processBio/clinvar/clinvar/variant_summary.txt",delimiter="\t")

summary

summary.GeneID.unique().shape

summary = summary[(summary.Type == "single nucleotide variant") & (summary.ReviewStatus.isin(["criteria provided, single submitter",
              "criteria provided, multiple submitters, no conflicts",
              "reviewed by expert panel",
              "practice guideline",
              ])) & (summary.Assembly == "GRCh38")]

handle = Entrez.efetch("protein", id="NP_060017", rettype="fasta", retmode="text")
lines = handle.read()
# res = json.loads(lines)
lines

summary.GeneID.unique().shape

summary.loc[7]

summary

missense = summary[(summary.ReviewStatus.isin(goodStatus)) & \
                   (summary.Type == "single nucleotide variant") & \
                   (summary.Name.str.contains("\(p.")) & \
                   (~summary.Name.str.contains("=")) & \
                   (summary.Assembly.str.contains("38"))]

missense

for each SNV, get the canonical sequence

from processBioDBs.utilities import getSequence

from easydict import EasyDict

from tqdm.notebook import tqdm

# sequences = EasyDict()

# for symbol in tqdm(missense.GeneSymbol.unique()):
#     if symbol not in sequences:
#         try:
#             sequences[symbol] = getSequence(symbol)
#         except:
#             continue

Dump to pickle file

import pickle

# with open("data/clinvar/seqs.pkl","wb") as f:
#     pickle.dump(sequences,f)

reload pickle (if didn't just process)

sequences = pickle.load(open("/data/projects/processBio/clinvar/clinvar/seqs.pkl","rb"))

1 letter to 3 letter AA conversion table

aatable = {"Ala":"A",
               "Arg": "R",
               "Asn": "N",
               "Asp": "D",
               "Cys": "C",
               "Gln": "Q",
               "Glu": "E",
               "Gly": "G",
               "His": "H",
               "Ile": "I",
               "Leu": "L",
               "Lys": "K",
               "Met": "M",
               "Phe": "F",
               "Pro": "P",
               "Ser": "S",
               "Thr": "T",
               "Trp": "W",
               "Tyr": "Y",
               "Val": "V"}

Get the variant from the name (expecting the form: \<original amino acid\>\<location\>\<new amino acid\>)

missense = missense.assign(variant=missense.Name.apply(lambda s: s[s.find("(p.") + 3:-1]))

missense = missense.assign(og=missense.variant.str.slice(0,3))
missense = missense.assign(loc=missense.variant.str.slice(3,-3))
missense = missense.assign(newAA=missense.variant.str.slice(-3,))

missense

drop rows where the original or new amino acid aren't valid

badRows = set()

for i,k in enumerate(missense.og):
    if k not in aatable:
         badRows.update([i])

for i,k in enumerate(missense.newAA):
    if k not in aatable:
        badRows.update([i])

missense = missense.drop(index=missense.index[list(badRows)])

make sure there are no duplicates

missense = missense.drop_duplicates(subset=["Name"])

missense

Find the rows where the canonical sequence I found match with the original amino acid given in the summary table

from tqdm.notebook import tqdm

import numpy as np

matches = np.zeros(missense.shape[0])
for i,(index,r) in tqdm(enumerate(missense.iterrows())):
    if r.GeneSymbol in sequences:
        s = sequences[r.GeneSymbol]
        loc = int(r["loc"]) - 1
        if loc >= len(s):
            continue
        if s[loc] == aatable[r.og]:
            matches[i] = 1
        

missense = missense.assign(match=matches)

missense

df = missense[missense.match == 1]

df = df.loc[(df['ClinicalSignificance'] == 'Likely benign')|
            (df['ClinicalSignificance'] == 'Benign')|
            (df['ClinicalSignificance'] == 'Pathogenic')|
            (df['ClinicalSignificance'] == 'Likely pathogenic')|
            (df['ClinicalSignificance'] == 'Benign/Likely benign')|
            (df['ClinicalSignificance'] == 'Pathogenic/Likely pathogenic')]

df = df.assign(label=(df.ClinicalSignificance.str.contains("pathogenic")) | \
               (df.ClinicalSignificance.str.contains("Pathogenic")))

df

df.groupby("GeneSymbol").count().quantile(.5)

df.groupby("GeneSymbol").label.unique().str.len().value_counts()

df

variants = []

for index,row in tqdm(df.iterrows()):
    s = sequences[row.GeneSymbol]
    loc = int(r["loc"]) - 1
    variant = s[:loc] + aatable[row["newAA"]] + s[loc+1:]
    variants.append(variant)

df = df.assign(sequence=variants)

df.to_csv("/ssdata/clinvar/clinvar/missenseVariants.csv")

df = pd.read_csv("/data/projects/processBio/clinvar/clinvar/missenseVariants.csv")

df

df.GeneID.unique().shape

ddf = pd.read_csv("/data/projects/processBio/clinvar/clinvar/clinvar.vep",delimiter="\t",header=92)

ddf

import mygene

mg = mygene.MyGeneInfo()

mg.querymany(["ENST00000330287"],returnall=True)
