==> 07_Gnomad.ipynb <==
import hail as hl

from processBioDBs.utilities import getSequence

mt = hl.import_vcf("/data/projects/processBio/gnomad/gnomad.exomes.r2.1.1.sites.vcf.gz",force_bgz=True)

mt.describe(widget=True)

res = hl.vep(mt,config="/home/dzeiberg/ensembl-vep/vepConfig.json")

missenseRes = res.filter_rows(res.vep.transcript_consequences.consequence_terms.contains(["missense_variant"]))



rows = missenseRes.row.take(4)

rows[0]

the following shows how to extract the variant, get the reference sequence and get the location of the mutation

rows[1].vep.transcript_consequences[0].amino_acids, getSequence(rows[1].vep.transcript_consequences[0].gene_symbol)[rows[1].vep.transcript_consequences[0].protein_start - 1]




def getVariedSequence(t):
    variant = t.amino_acids
    loc = t.protein_start - 1
    seq = "ABC"
    return seq[:loc] + variant[-1] + seq[loc+1:]

variants = missenseRes.vep.transcript_consequences.map(lambda x: getVariedSequence(x))
