==> 08_Yeast_Compliment_Assay_Data.ipynb <==
aaTable = {"Ala":"A",
           "Arg": "R",
           "Asn": "N",
           "Asp": "D",
           "Cys": "C",
           "Gln": "Q",
           "Glu": "E",
           "Gly": "G",
           "His": "H",
           "Ile": "I",
           "Leu": "L",
           "Lys": "K",
           "Met": "M",
           "Phe": "F",
           "Pro": "P",
           "Ser": "S",
           "Thr": "T",
           "Trp": "W",
           "Tyr": "Y",
           "Val": "V"}

import pandas as pd
import torch
import esm
from tqdm.notebook import tqdm

from processBioDBs.utilities import getSequence

# Load ESM-1b model
model, alphabet = esm.pretrained.esm1b_t33_650M_UR50S()
batch_converter = alphabet.get_batch_converter()


model = model.to("cuda:1")

def getSequenceRepresentation(Data):
    batch_labels, batch_strs, batch_tokens = batch_converter(Data)
    batch_tokens = batch_tokens.to("cuda:1")
    # Extract per-residue representations (on CPU)
    with torch.no_grad():
        results = model(batch_tokens, repr_layers=[33], return_contacts=True)
    token_representations = results["representations"][33].cpu()
    del results, batch_labels, batch_strs, batch_tokens
    # Generate per-sequence representations via averaging
    # NOTE: token 0 is always a beginning-of-sequence token, so the first residue is token 1.
    sequence_representations = []
    for i, (_, seq) in enumerate(Data):
        sequence_representations.append(token_representations[i, 1 : len(seq) + 1].cpu().numpy())
    del token_representations
    return sequence_representations

def prepSeq(s,loc,windowSize=510):
    sPrime = s[max(0, loc-windowSize) : min(len(s), loc + windowSize)]
    return sPrime

# PTEN

pten = pd.read_csv("/data/projects/processBio/ycaData/parse_pten.csv")

pten=pten.assign(variant=pten.variant.apply(lambda s: (aaTable[s[:3]], int(s[3:-3]) - 1, aaTable[s[-3:]])))

ptenSeq = getSequence("pten")

variants = []
for idx,row in pten.iterrows():
    og,loc,var = row["variant"]
    if ptenSeq[loc] != og:
        variants.append("")
        continue
    variants.append(ptenSeq[:loc] + var + ptenSeq[loc+1:])

pten = pten.assign(varSeqe=variants)

representations = []
batchSize=1
for idx,row in tqdm(pten.iterrows(),total=pten.shape[0]):
    rep = getSequenceRepresentation([(idx, prepSeq(row.varSeqe, row.variant[1]))])
    representations.append(rep)
#     try:
#         rep = getSequenceRepresentation([(idx,prepSeq(row.varSeqe, row.variant[1]))])
#         representations.append(rep)
#     except:
#         representations.append([])

pten = pten.assign(representation=representations)

pten.to_pickle("/data/projects/processBio/ycaData/parse_pten.pkl")

# CALM1

calm = pd.read_csv("/data/projects/processBio/ycaData/parse_calm1.csv")

calm=calm.assign(variant=calm.variant.apply(lambda s: (aaTable[s[:3]], int(s[3:-3]) - 1, aaTable[s[-3:]])))

calmSeq = getSequence("calm1")

variants = []
for idx,row in calm.iterrows():
    og,loc,var = row["variant"]
    if calmSeq[loc] != og:
        variants.append("")
        continue
    variants.append(calmSeq[:loc] + var + calmSeq[loc+1:])

calm = calm.assign(varSeqe=variants)

representations = []
batchSize=1
for idx,row in tqdm(calm.iterrows(),total=calm.shape[0]):
    rep = getSequenceRepresentation([(idx, prepSeq(row.varSeqe, row.variant[1]))])
    representations.append(rep)


calm = calm.assign(representation=representations)

calm.to_pickle("/data/projects/processBio/ycaData/parse_calm1.pkl")

# SUMO

sumo = pd.read_csv("/data/projects/processBio/ycaData/parse_sumo.csv")
sumo = sumo.assign(variant=sumo.variant.apply(lambda s: (aaTable[s[:3]], int(s[3:-3]) - 1, aaTable[s[-3:]])))

sumoSeq = getSequence("sumo1")

variants = []
for idx,row in sumo.iterrows():
    og,loc,var = row["variant"]
    if sumoSeq[loc] != og:
        variants.append("")
        continue
    variants.append(sumoSeq[:loc] + var + sumoSeq[loc+1:])

sumo = sumo.assign(varSeqe=variants)

representations = []
batchSize=1
for idx,row in tqdm(sumo.iterrows(),total=sumo.shape[0]):
    rep = getSequenceRepresentation([(idx, prepSeq(row.varSeqe, row.variant[1]))])
    representations.append(rep)



sumo = sumo.assign(representation=representations)

sumo.to_pickle("/data/projects/processBio/ycaData/parse_sumo.pkl")

# tpmt

tpmt = pd.read_csv("/data/projects/processBio/ycaData/parse_tpmt.csv")
tpmt = tpmt.assign(variant=tpmt.variant.apply(lambda s: (aaTable[s[:3]], int(s[3:-3]) - 1, aaTable[s[-3:]])))

tpmtSeq = getSequence("tpmt")

variants = []
for idx,row in tpmt.iterrows():
    og,loc,var = row["variant"]
    if tpmtSeq[loc] != og:
        variants.append("")
        continue
    variants.append(tpmtSeq[:loc] + var + tpmtSeq[loc+1:])

tpmt = tpmt.assign(varSeqe=variants)

representations = []
batchSize=1
for idx,row in tqdm(tpmt.iterrows(),total=tpmt.shape[0]):
    rep = getSequenceRepresentation([(idx, prepSeq(row.varSeqe, row.variant[1]))])
    representations.append(rep)




tpmt = tpmt.assign(representation=representations)

tpmt.to_pickle("/data/projects/processBio/ycaData/parse_tpmt.pkl")

def getRep(rep,L,originalWindow=510, W=3):
    i,j = max(0, L-originalWindow),min(len(rep), L + originalWindow)
    k,m = max(0,L - W - i), min(len(rep),L + W - i + 1)
    return rep[k : m]

tpmt.head()

import numpy as np

def prepSeq(row, windowSizes=[2,4,8,16,32,64,128,256]):
    rep = np.array(row.representation[0])
    L = row.variant[1]
    reps = [getRep(rep,L, originalWindow=510,W=w).mean(0) for w in windowSizes]
    xi = np.concatenate(reps)
    return xi

tpmt = tpmt.assign(xi=tpmt.apply(lambda row: prepSeq(row),axis=1))

tpmt.to_pickle("/data/projects/processBio/ycaData/parse_tpmt.pkl")

sumo = sumo.assign(xi=sumo.apply(lambda row: prepSeq(row),axis=1))

sumo.to_pickle("/data/projects/processBio/ycaData/parse_sumo.pkl")

calm = calm.assign(xi=calm.apply(lambda row: prepSeq(row),axis=1))

calm.to_pickle("/data/projects/processBio/ycaData/parse_calm1.pkl")

pten = pten.assign(xi=pten.apply(lambda row: prepSeq(row),axis=1))

pten.to_pickle("/data/projects/processBio/ycaData/parse_pten.pkl")

pten = pd.read_pickle("/data/projects/processBio/ycaData/parse_pten.pkl")

# GnomAD Data

from glob import glob

gnomadFiles = glob("/data/projects/processBio/gnomad/yca/*.csv")

gnomadDFs = []

def makeVariedSequence(row,seq):
    og,loc,var= row.variant
    if loc == -1 or seq[loc] != og:
        return ""
    return seq[:loc] + var + seq[loc+1:]

def parseConsequence(c):
    try:
        og = aaTable[c[2:5]]
    except KeyError:
        return ("X",-1,"X")
    try:
        loc = int(c[5:-3]) - 1
    except ValueError:
        return ("X",-1,"X")
    try:
        var = aaTable[c[-3:]]
    except KeyError:
        return ("X",-1,"X")
    return (og,loc,var)

for file in gnomadFiles:
    df = pd.read_csv(file)
    geneSymbol = file[file.rfind("/") + 1 : -4]
    seq = getSequence(geneSymbol)
    df = df.assign(variant=df["HGVS Consequence"].apply(lambda s: parseConsequence(s)))
    df = df.assign(seq = df.apply(lambda row: makeVariedSequence(row,seq),axis=1))
    df = df.assign(embedding=df.seq.apply(lambda s: getSequenceRepresentation([(-1, s)])))
    gnomadDFs.append(df)

gnomadDFs[0]

for name, df in zip(gnomadFiles, gnomadDFs):
    df.to_pickle(name.replace(".csv",".pkl"))

# Generate vectors for gnomadFiles

gnomadDFs = []
gnomadFileNames = []
for file in glob("/data/projects/processBio/gnomad/yca/*.pkl"):
    gnomadDFs.append(pd.read_pickle(file))
    gnomadFileNames.append(file)

from processBioDBs.utilities import prepSeq as convertEmbeddingMatrixToVector

import numpy as np

for i, dfi in enumerate(gnomadDFs):
    gnomadDFs[i] = dfi.assign(xi=dfi.apply(lambda row: convertEmbeddingMatrixToVector(row["embedding"][0],
                                                                                     row.variant[1]),axis=1))

gnomadFileNames[0]

gnomadDFs[0]

for dfi,fn in zip(gnomadDFs, gnomadFileNames):
    dfi.to_pickle(fn)
