==> 09_Summary.ipynb <==
import pandas as pd
import numpy as np

hgmd = pd.read_csv("/data/projects/processBio/hgmd/variant_df.csv")

hgmd = hgmd.rename({"GENE":"GeneSymbol"},axis=1)

hgmd = hgmd.assign(label=True)

hgmdX = np.load("/data/projects/processBio/hgmd/X.npy")

clinvarDF = pd.read_csv("/data/projects/processBio/clinvar/clinvar/parseClinvarWithVariant.csv").rename({"variantSeq": "seq"},axis=1)

clinvarDF[clinvarDF.seq != ""]

import pickle
clinvarEmbeddings = pickle.load(open("/data/projects/processBio/clinvar/clinvar/parseClinvarRepresentations.pkl","rb"))

clinvarDF = clinvarDF[[len(e) > 0 for e in clinvarEmbeddings]]

clinvarDF

clinvarX = np.load("/data/projects/processBio/clinvar/clinvar/X.npy")

clinvarX.shape

gnomad = pd.read_pickle("/data/projects/processBio/gnomad/gnomad.missenseVariants.pd.pkl")

gnomad = gnomad[gnomad.apply(lambda row: len(row.seq) > int(row.Protein_position) - 1,axis=1)]

gnomadX = np.load("/data/projects/processBio/gnomad/gnomadValidMissenseVariantEmbeddings.npy")

gnomadX.shape

gnomadSymbols = gnomad.INFO.apply(lambda i: i["SYMBOL"] if "SYMBOL" in i else "")
gnomad=gnomad.assign(GeneSymbol=gnomadSymbols)

gnomad[gnomad.GeneSymbol == ""]

clinvarSymbols = clinvarDF["GeneSymbol"]

hgmdSymbols = hgmd["GeneSymbol"]

symbols = pd.concat([clinvarSymbols,
                     hgmdSymbols,
                     gnomadSymbols])

symbols.shape

symbols.value_counts().shape, "{:.3f}".format(symbols.value_counts().mean())

# Combine all datasets

gnomad = gnomad.assign(label=np.nan)

data =pd.concat((gnomad[["seq","label","GeneSymbol"]],
           clinvarDF[["seq","label","GeneSymbol"]],
           hgmd[["label", "seq","GeneSymbol"]]))

X = np.concatenate((gnomadX, clinvarX,hgmdX))

gnomadX.shape, clinvarX.shape, hgmdX.shape, X.shape

clinvarDF.label.value_counts()

np.save("/data/projects/processBio/X.npy",X)

data.to_pickle("/data/projects/processBio/df.pd.pkl")

alphas = []
numPos = []
numNeg = []
for (idx,group) in data[~data.label.isna()].groupby("GeneSymbol"):
    labels = group.label
    if labels.sum() > 0 and labels.sum() < labels.shape[0]:
        alphas.append(labels.sum() / labels.shape[0])
        numPos.append(labels.sum())
        numNeg.append(labels.shape[0] - labels.sum())

len(alphas)

(data.GeneSymbol == "").any()

data.groupby("GeneSymbol").count()

import matplotlib.pyplot as plt
plt.hist(alphas,)

# Can I skip clustering by using the mutpred data clusters?

mutpreddf = pd.read_csv("/data/projects/mutpred/mutpred2_revised_val_cleaned_cluster50.txt",header=10,delimiter="\t")
mutpreddf = mutpreddf.assign(symbol=mutpreddf.internal_id.apply(lambda x: [xi.replace("_HUMAN","") for xi in x.split(";")  if "HUMAN" in xi]))

mutpreddf

mpsymbols = set([s[0] for s in mutpreddf.symbol if len(s)])

pd.Series(data.GeneSymbol.unique()).isin(mpsymbols).value_counts()

I cannot...

data[~data.label.isna()].groupby("GeneSymbol").count().mean()

data[~data.label.isna()]

data.label.value_counts()

data.label.shape
