==> 13_CDHIT.ipynb <==
import pandas as pd

df = pd.read_pickle("/data/projects/processBio/df.pd.pkl")

df = df[~df.label.isna()]

df

symbols = df.GeneSymbol.unique()

len(symbols)

from processBioDBs.utilities import getSequence
from tqdm.notebook import tqdm

import ensembl_rest

seqs = pickle.load(open("/data/projects/processBio/clinvar/clinvar/referenceSeqs.pkl","rb"))

idMap = {}
with open('/data/projects/processBio/df.fasta',"w") as f:
    for i, (k,v) in enumerate(seqs.items()):
        f.write(">{}|{} {}\n{}\n".format(i,k,k,v))
        idMap[i] = k

pickle.dump(idMap,open("/data/projects/processBio/idxToSymbolMap.pkl",'wb'))

!head /data/projects/processBio/df.fasta

!cat /data/projects/processBio/dfCDHITRESULTS04.clstr | grep Cluster | wc -l

len(seqs)
