==> 14_Trust_The_Reprocess.ipynb <==
from IPython.core.display import display, HTML
display(HTML("<style>.container { width:100% !important; }</style>"))


import numpy as np
import pandas as pd
from tqdm.notebook import tqdm

aaTable = {"Ala":"A",
           "Arg": "R",
           "Asn": "N",
           "Asp": "D",
           "Cys": "C",
           "Gln": "Q",
           "Glu": "E",
           "Gly": "G",
           "His": "H",
           "Ile": "I",
           "Leu": "L",
           "Lys": "K",
           "Met": "M",
           "Phe": "F",
           "Pro": "P",
           "Ser": "S",
           "Thr": "T",
           "Trp": "W",
           "Tyr": "Y",
           "Val": "V"}

import pickle

npid2seq = pickle.load(open("/data/projects/processBio/ncbi/npid2seq.pkl","rb"))
# npid2seq = pickle.load(open("/ssdata/projects"))

# gnomAD

df = pd.read_csv("/data/projects/processBio/gnomad/gnomad.exomes.r2.1.1.sites.vcf.vep",
                 header=93,delimiter="\t")

df

Originally 103,321,415 variants

df = df[df.Consequence.str.contains("missense")]

df

1,652,025 missense variants

df = df.assign(info=df.Extra.apply(lambda s: dict([l.split("=") for l in s.split(";")])))

df.Feature.unique().shape

from ensembl_rest import sequence_id, lookup, symbol_lookup, HTTPError
def getSequenceFromTranscript(transcriptID):
    return sequence_id(lookup(transcriptID,
                              params={"expand":True})["Translation"]["id"])["seq"]

transcripts = df.Feature.unique()

gnomadTranscriptToSeq = {}
for t in tqdm(transcripts):
    gnomadTranscriptToSeq[t] = getSequenceFromTranscript(t)

import pickle

pickle.dump(gnomadTranscriptToSeq, open("/data/projects/processBio/gnomad/transcriptToSeq.pkl","wb"))

gnomadTranscriptToSeq = pickle.load(open("/data/projects/processBio/gnomad/transcriptToSeq.pkl","rb"))

def makeGnomADSeq(row):
    og,var = row.Amino_acids[0], row.Amino_acids[-1]
    loc = int(row["Protein_position"]) - 1
    refSeq= gnomadTranscriptToSeq[row["Feature"]]
    if len(refSeq) <= loc:
        print("invalid seq len")
        return ""
    if refSeq[loc] == og:
        return refSeq[:loc] + var + refSeq[loc+1:]
    print("invalid reference AA")
    return ""

df = df.assign(seq=df.apply(lambda row: makeGnomADSeq(row),axis=1))

df[df.seq != ""]

df.to_pickle("/data/projects/processBio/gnomad/df.reprocessed.pd.pkl")

df

# HGMD

hgmd = pd.read_csv("/data/projects/processBio/hgmd/HGMD_PRO_2019_1_hg19.vcf",delimiter="\t",header=14)

hgmd = hgmd.assign(INFO = hgmd.INFO.apply(lambda s: dict([l.split("=") for l in s.split(";")])))

hgmd

Originally 229,161 HGMD variants

hgmd = hgmd[hgmd.INFO.apply(lambda i: i["CLASS"] == "DM")]

hgmd

180,092 variants with class DM

from Bio import Entrez
Entrez.email = 'zeiberg.d2@northeastern.edu'

def getSequenceFromNPID(npid):
    "Return the protein sequence from "
    handle = Entrez.efetch(db="protein",id=npid, rettype="fasta", retmode="text")
    lines = handle.readlines()
    lines = [l.strip() for l in lines]
    return "".join(lines[1:])

def getProt(i):
    if "PROT" in i:
        return i["PROT"][:i["PROT"].find(":")]
    return ""

def hasProt(i):
    if "PROT" in i:
        return 1
    return 0

def getHGMDVariantInfo(i):
    if "PROT" not in i:
        return (np.nan, np.nan)
    npid, variant = i["PROT"].split(":")
    variant = variant[2:]
    og,loc,var = variant[0], variant[1:-1], variant[-1]
    if og not in aaTable.values() or var not in aaTable.values():
        return (np.nan, np.nan)
    try:
        loc = int(loc) - 1
    except ValueError:
        return (np.nan, np.nan)
    return (npid, (og,loc,var))

hgmdVariantInfo = hgmd.INFO.apply(lambda i: getHGMDVariantInfo(i))

hgmd = hgmd.assign(variantInfo = hgmdVariantInfo[hgmdVariantInfo.apply(lambda t: not pd.isna(t[0]))])

hgmd

hgmd[~hgmd.variantInfo.isna()]

def getHGMDNPIDs(info):
    if not pd.isna(info):
        npid,_ = info
        if npid not in npid2seq:
            npid2seq[npid] = getSequenceFromNPID(npid)
    

for info in tqdm(hgmd.variantInfo):
    getHGMDNPIDs(info)

def makeHGMDSeq(info):
    if pd.isna(info):
        return ""
    npid, (og,loc,var) = info
    refSeq = npid2seq[npid]
    if len(refSeq) < loc:
        return ""
    if refSeq[loc] != og:
        return ""
    return refSeq[:loc] + var + refSeq[loc+1:]

hgmdSeqs = []
for info in tqdm(hgmd.variantInfo):
    hgmdSeqs.append(makeHGMDSeq(info))

hgmd = hgmd.assign(seq=hgmdSeqs)

hgmd[hgmd.seq != ""]

# Clinvar

summary = pd.read_csv("/data/projects/processBio/clinvar/clinvar/variant_summary.txt",delimiter="\t")


summary = summary[(summary.Type == "single nucleotide variant") & (summary.ReviewStatus.isin(["criteria provided, single submitter",
              "criteria provided, multiple submitters, no conflicts",
              "reviewed by expert panel",
              "practice guideline",
              ])) & (summary.Assembly == "GRCh38")]

summary = summary.assign(nmid=summary.Name.apply(lambda n: n[:n.find("(")]))

summary

767,700 single nucleotide variants with reference genome GRCh38 with one-star review status

g2rs = pd.read_csv("/data/projects/processBio/ncbi/gene2refseq",delimiter="\t")

g2rs = g2rs[(g2rs.status == "REVIEWED") 
            & (g2rs.assembly == "Reference GRCh38.p13 Primary Assembly")
            & (g2rs["protein_accession.version"] != "-")]

nmid2npid = {}
for nmid in tqdm(summary.nmid.unique()):
    nmid2npid[nmid] = g2rs[g2rs["RNA_nucleotide_accession.version"] == nmid]["protein_accession.version"].unique()

summary = summary.assign(npid=summary.nmid.apply(lambda n: nmid2npid[n]))


for npid in tqdm(set([s[0] for s in summary.npid.values if len(s)])):
    if npid not in npid2seq:
        try:
            npid2seq[npid] = getSequenceFromNPID(npid)
        except Entrez.HTTPError:
            raise

pickle.dump(npid2seq,open("/data/projects/processBio/ncbi/npid2seq.pkl","wb"))

summary

summary = summary.assign(npid=summary[summary.npid.str.len() != 0].npid.apply(lambda n: n[0] if len(n) else np.nan))

summary[~summary.npid.isna()]

623,532 variants found a matching protein id

summary[~summary.npid.isna()].ClinicalSignificance.value_counts()

clinvar = summary[(~summary.npid.isna()) & (summary.ClinicalSignificance.isin(["Likely benign",
                                                                     "Benign", 
                                                                     "Pathogenic",
                                                                     "Likely pathogenic",
                                                                     "Benign/Likely benign",
                                                                     "Pathogenic/Likely pathogenic"]))]

clinvar.loc[33]

328,243 variants have matching protein and valid status

def extractVariantFromClinvar(n):
    if n.rfind("(p.") != -1:
        variant = n[n.rfind("(p.") + 3 : -1]
#         print(variant)
        og,loc,var = variant[:3], variant[3:-3], variant[-3:]
        try:
            loc = int(loc) -1
        except ValueError:
            return np.nan, np.nan,np.nan
        if variant[-1] == "=":
#             print("variant is =")
            return np.nan, np.nan,np.nan
        return og,loc,var
    return np.nan, np.nan, np.nan

clinvar = clinvar.assign(variantInfo=clinvar.Name.apply(lambda n: extractVariantFromClinvar(n)))

# clinvar = clinvar[clinvar.variantInfo.apply(lambda t: not pd.isna(t[0]))]
clinvar[clinvar.variantInfo.apply(lambda t: pd.isna(t[0]))].RCVaccession

clinvar

73,371 variants have matching protein, valid status, and have alternate residue != "="

clinvar.ClinicalSignificance.value_counts()

clinvar

def makeClinvarSeq(row):
    npid = row["npid"]
    refSeq = npid2seq[npid]
    og,loc,var = row["variantInfo"]
    if og not in aaTable or var not in aaTable:
        if var != "Ter":
            print('invalid parsing of variant... ', og, loc, var)
        return ""
    og = aaTable[og]
    var = aaTable[var]
    if len(refSeq) <= loc:
        print('invalid seq len')
        return ""
    if refSeq[loc] != og:
        print('invalid ref AA')
        return ""
    return refSeq[:loc] + var + refSeq[loc+1:]

clinvar = clinvar.assign(seq = clinvar.apply(lambda row: makeClinvarSeq(row),axis=1))

clinvar[clinvar.seq != ""]

Successfully retreived 55,087 variants

# Combine Labeled Data

hgmd = hgmd.assign(ClinSigSimple=1)
hgmd = hgmd.assign(db="hgmd")

clinvar = clinvar.assign(db="clinvar")

hgmd = hgmd.assign(npid = hgmd.variantInfo.apply(lambda i: i[0] if not pd.isna(i) else np.nan))

labeledDF = pd.concat((hgmd[hgmd.seq != ""][["ClinSigSimple", "seq", "npid", "db","variantInfo"]],
  clinvar[clinvar.seq != ""][["npid", "seq", "ClinSigSimple", "db","variantInfo"]]))

labeledDF

labeledDF.npid.unique().shape

labeledDF.ClinSigSimple.value_counts()

labeledDF.groupby("npid").count().seq.mean()

(labeledDF.groupby("npid").count() > 100).sum()

import torch
import esm
import os
import torch.nn as nn
os.environ['CUDA_DEVICE_ORDER']='PCI_BUS_ID'
os.environ['CUDA_VISIBLE_DEVICES']='1,2,3'


model, alphabet = esm.pretrained.esm1b_t33_650M_UR50S()
batch_converter = alphabet.get_batch_converter()

model = model.to("cuda:2")

from tqdm.notebook import trange

def embedSeq(Data):
    BATCHSIZE=1
    representations = []
    for start in trange(0,len(Data),BATCHSIZE):
        batch_labels, batch_strs, batch_tokens = batch_converter(Data[start : start + BATCHSIZE])
        # Extract per-residue representations (on CPU)
        with torch.no_grad():
            results = model(batch_tokens.to("cuda:2"), repr_layers=[33], return_contacts=True)
        token_representations = results["representations"][33].cpu()
        del results, batch_labels, batch_strs, batch_tokens
        representations.append(token_representations[0,1:-1].cpu().numpy())
        del token_representations
    return representations

def prepSeq(row):
    if type(row["variantInfo"][1]) == tuple:
        loc= row["variantInfo"][1][1]
    else:
        loc = row["variantInfo"][1]
    seq = row["seq"]
    return seq[max(0,loc-510) : min(len(seq), loc + 510)]

labeledDF = labeledDF.assign(preppedSeq = labeledDF.apply(lambda row: prepSeq(row),axis=1))

df.head()

def prepGnomad(row):
    loc = int(row["Protein_position"])-1
    seq = row["seq"]
    return seq[max(0,loc-510) : min(len(seq), loc + 510)]

df= df.assign(preppedSeq=df.apply(lambda row: prepGnomad(row),axis=1))

from tqdm.notebook import trange

GnomadData = list(df.preppedSeq.items())

gnomadEmbeddings = embedSeq(GnomadData)

df = df.assign(embedding=gnomadEmbeddings)

df.to_pickle('/data/projects/processBio/gnomadReprocessed.pkl')


Data = list(labeledDF.preppedSeq.items())

embeddings = embedSeq(Data)

embeddings

labeledDF = labeledDF.assign(embedding=embeddings)

labeledDF

import pickle

labeledDF = labeledDF.assign(vec=labeledDF.embedding.apply(lambda e: e.mean(0)))

labeledDF.to_pickle("/data/projects/processBio/labeledDFReprocessed.pkl")

labeledDF.ClinSigSimple.value_counts()

labeledDF.groupby("npid").ClinSigSimple.value_counts().describe()

ls /data/projects/processBio/
